app/rq_api/pyproject.toml
[project]
name = "sandbox"
version = "0.1.0"
description = "Sandbox API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-libcloud>=3.9.0",
    "cryptography>=46.0.5",
    "fastapi",
    "uvicorn",
    "rq",
    "redis",
    "apache-libcloud",
]
